+++ acceptance-tests/cosmosdb/cosmosdb_test.go
@@ -3,6 +3,7 @@ package cosmosdb_test
 import (
 	"acceptancetests/apps"
 	"acceptancetests/helpers"
+	"acceptancetests/helpers/matchers"
 	"acceptancetests/helpers/random"
 	"fmt"
 
@@ -34,7 +35,7 @@ var _ = Describe("CosmosDB", func() {
 		helpers.AppStart(appOne, appTwo)
 
 		By("checking that the app environment has a credhub reference for credentials")
-		Expect(binding.Credential()).To(helpers.HaveCredHubRef)
+		Expect(binding.Credential()).To(matchers.HaveCredHubRef)
 
 		By("checking that the specified database has been created")
 		databases := appOne.GET("/")
