+++ acceptance-tests/mssql/mssql_failover_group_test.go
@@ -3,6 +3,7 @@ package mssql_test
 import (
 	"acceptancetests/apps"
 	"acceptancetests/helpers"
+	"acceptancetests/helpers/matchers"
 	"acceptancetests/helpers/random"
 
 	. "github.com/onsi/ginkgo"
@@ -28,7 +29,7 @@ var _ = Describe("MSSQL Failover Group", func() {
 		helpers.AppStart(appOne, appTwo)
 
 		By("checking that the app environment has a credhub reference for credentials")
-		Expect(bindingOne.Credential()).To(helpers.HaveCredHubRef)
+		Expect(bindingOne.Credential()).To(matchers.HaveCredHubRef)
 
 		By("creating a schema using the first app")
 		schema := random.Name(random.WithMaxLength(10))
